requirements.txt
gradio==5.1.*
boto3
python-dotenv==1.0.1